+++ example/publisher.cc
@@ -41,33 +41,33 @@ void signal_handler(int _signal)
 int main(int argc, char **argv)
 {
   // Install a signal handler for SIGINT and SIGTERM.
-  std::signal(SIGINT,  signal_handler);
-  std::signal(SIGTERM, signal_handler);
+  // std::signal(SIGINT,  signal_handler);
+  // std::signal(SIGTERM, signal_handler);
 
   // Create a transport node and advertise a topic.
   ignition::transport::Node node;
-  std::string topic = "/foo";
+  // std::string topic = "/foo";
 
-  auto pub = node.Advertise<ignition::msgs::StringMsg>(topic);
-  if (!pub)
-  {
-    std::cerr << "Error advertising topic [" << topic << "]" << std::endl;
-    return -1;
-  }
+  // auto pub = node.Advertise<ignition::msgs::StringMsg>(topic);
+  // if (!pub)
+  // {
+  //   std::cerr << "Error advertising topic [" << topic << "]" << std::endl;
+  //   return -1;
+  // }
 
   // Prepare the message.
-  ignition::msgs::StringMsg msg;
-  msg.set_data("HELLO");
+  // ignition::msgs::StringMsg msg;
+  // msg.set_data("HELLO");
 
   // Publish messages at 1Hz.
-  while (!g_terminatePub)
-  {
-    if (!pub.Publish(msg))
-      break;
+  // while (!g_terminatePub)
+  // {
+  //   if (!pub.Publish(msg))
+  //     break;
 
-    std::cout << "Publishing hello on topic [" << topic << "]" << std::endl;
-    std::this_thread::sleep_for(std::chrono::milliseconds(1000));
-  }
+  //   std::cout << "Publishing hello on topic [" << topic << "]" << std::endl;
+  //   std::this_thread::sleep_for(std::chrono::milliseconds(1000));
+  //  }
 
   return 0;
 }